Rename connection module to sync and ConnectionManager to SyncManager
