--- p2panda-net/src/connection/mod.rs
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod handler;
-mod manager;
-mod sync;
-
-pub use handler::{SyncConnection, SYNC_CONNECTION_ALPN};
-pub use manager::ConnectionManager;
