+++ p2panda-net/src/network.rs
@@ -24,11 +24,11 @@ use tracing::{debug, error, error_span, warn, Instrument};
 
 use crate::addrs::DEFAULT_STUN_PORT;
 use crate::config::{Config, DEFAULT_BIND_PORT};
-use crate::connection::SYNC_CONNECTION_ALPN;
 use crate::discovery::{Discovery, DiscoveryMap};
 use crate::engine::Engine;
 use crate::handshake::{Handshake, HANDSHAKE_ALPN};
 use crate::protocols::{ProtocolHandler, ProtocolMap};
+use crate::sync::SYNC_CONNECTION_ALPN;
 use crate::{JoinErrToStr, NetworkId, RelayUrl, TopicId};
 
 /// Maximum number of streams accepted on a QUIC connection.
