+++ p2panda-net/src/sync/mod.rs
@@ -1,5 +1,11 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod handler;
+mod manager;
+
+pub use handler::{SyncConnection, SYNC_CONNECTION_ALPN};
+pub use manager::SyncManager;
+
 use std::sync::Arc;
 
 use anyhow::Result;
